labs/lab-05: Add checker infrastructure
1. Added 3 simple tests that check for the right quotient
and remainder.

2. Added 3 simple tests that check for the right product of
the multiplication of the given numbers.

3. Added 4 tests for the word and dword vectors, for the sum of
squares and for the 64 bits sum of squares. Furthermore, I added
the task in the `README.md` file for the latter. In addition,
with a view to make the testing easier, all the tasks should
be implemented in the file `sum_array`, as the implementation of
the tasks in separate files was not clearly mentioned.

4. Added 2 tests for both of the sums. The existence of the
second binary is checked before executing it, as the creation
of that file is stated in the README.md.

5. Added 4 tests for positive, negative, even and odd numbers.
To make the testing simpler, the array that should be declared
is written in the README.md file.

Fixes #38

Signed-off-by: Andrei Rusanescu <andreirusanescu154@gmail.com>